reverts individual hook change
